requirements.txt
Flask==2.3.3
face-recognition==1.3.0
opencv-python==4.8.1.78
pymongo==4.6.0
python-dotenv==1.0.0
Pillow==10.0.1
numpy==1.24.3
requests==2.31.0
gunicorn==21.2.0
